feat: favorite ui app bar; fix: room db config

+++ app/src/main/java/com/vau/studio/iosstyle/idialer_phone/core/DbUtils.kt
@@ -1,7 +1,7 @@
 package com.vau.studio.iosstyle.idialer_phone.core
 
 object DbUtils {
-    fun isSuccess(operandNumber: Long) : Boolean {
-        return operandNumber.toInt() == 0
+    fun isSuccess(operandNumber: Int) : Boolean {
+        return operandNumber == 0
     }
 }
